test_requirements.txt
Flask
PyYAML
nose
coverage
diff-cover
pylint
pep8
mock
coveralls
